changed folder name and added count of exercice in a set
